requirements.txt
requests==2.31.0
python-dotenv==1.0.0
groq>=0.9.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
